update log aop, done api create FC, reformat project
