+++ src/main/java/com/luke/fcmanagement/advice/ApiExceptionHandler.java
@@ -1,9 +1,9 @@
 package com.luke.fcmanagement.advice;
 
 import com.luke.fcmanagement.constants.ErrorCode;
+import com.luke.fcmanagement.exception.BusinessException;
 import com.luke.fcmanagement.model.ApiError;
 import com.luke.fcmanagement.model.ApiResponse;
-import com.luke.fcmanagement.model.ErrorMsg;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindException;
@@ -12,7 +12,6 @@ import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.context.request.WebRequest;
 
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -24,21 +23,25 @@ public class ApiExceptionHandler {
      */
     @ExceptionHandler(Exception.class)
     @ResponseStatus(value = HttpStatus.INTERNAL_SERVER_ERROR)
-    public ApiResponse handleAllException(Exception ex, WebRequest request) {
+    public ApiResponse handleAllException(Exception ex) {
         String msg = StringUtils.isBlank(ex.getLocalizedMessage()) ? ex.toString() : ex.getLocalizedMessage();
-        ErrorMsg errorMessage = new ErrorMsg(msg, null);
-        ApiError apiError = new ApiError(errorMessage);
+        ApiError apiError = new ApiError(msg, null);
         return ApiResponse.fail(ErrorCode.INTERNAL_ERROR, apiError);
     }
 
     @ExceptionHandler(BindException.class)
     @ResponseStatus(value = HttpStatus.BAD_REQUEST)
-    public ApiResponse handleBindException(BindException ex, WebRequest request) {
+    public ApiResponse handleBindException(BindException ex) {
         BindingResult bindingResult = ex.getBindingResult();
         Map<String, Object> err = bindingResult.getFieldErrors().stream().collect(Collectors.toMap(FieldError::getField, FieldError::getDefaultMessage));
-        ErrorMsg errorMessage = new ErrorMsg(ErrorCode.VALIDATE_FAIL.getMessage(), err);
-        ApiError apiError = new ApiError(errorMessage);
+        ApiError apiError = new ApiError(ErrorCode.VALIDATE_FAIL.getMessage(), err);
         return ApiResponse.fail(ErrorCode.VALIDATE_FAIL, apiError);
+    }
+
+    @ExceptionHandler({BusinessException.class})
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ApiResponse handleBusinessException(BusinessException ex) {
+        return ApiResponse.fail(ex.getErrorCode());
     }
 
 }
