+++ src/main/java/com/luke/fcmanagement/model/ApiError.java
@@ -5,10 +5,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.HashMap;
+import java.util.Map;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class ApiError {
-    private ErrorMsg errorMsg;
+    private String message;
+    Map<String, Object> errors = new HashMap<>();
 }
