--- src/main/java/com/luke/fcmanagement/model/ErrorMsg.java
@@ -1,19 +1,0 @@
-package com.luke.fcmanagement.model;
-
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class ErrorMsg {
-    private String message;
-    Map<String, Object> errors = new HashMap<>();
-}
