+++ src/main/java/com/luke/fcmanagement/module/file/FileSaver.java
@@ -1,0 +1,7 @@
+package com.luke.fcmanagement.module.file;
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface FileSaver {
+    String saveFile(MultipartFile file, String folderPath, String fileName);
+}
