--- src/main/java/com/luke/fcmanagement/utils/file_utils/FileSaver.java
@@ -1,7 +1,0 @@
-package com.luke.fcmanagement.utils.file_utils;
-
-import org.springframework.web.multipart.MultipartFile;
-
-public interface FileSaver {
-    String saveFile(MultipartFile file, String folderPath, String fileName);
-}
