--- src/main/java/com/luke/fcmanagement/validate/validator/Size.java
@@ -1,21 +1,0 @@
-package com.luke.fcmanagement.validate.validator;
-
-import com.luke.fcmanagement.validate.annotation.MultipartArraySizeValidator;
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target({ElementType.FIELD, ElementType.PARAMETER})
-@Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = MultipartArraySizeValidator.class)
-public @interface Size {
-    int min() default 0;
-    int max() default Integer.MAX_VALUE;
-    String message() default "Invalid size";
-    Class<?>[] groups() default {};
-    Class<? extends Payload>[] payload() default {};
-}
